[Enhancement](memory) Move concurrent fastutil maps to fe-foundation module
Address reviewer feedback (morrySnow): move ConcurrentLong2ObjectHashMap
and ConcurrentLong2LongHashMap from fe-core/common to fe-foundation/util.
Add fastutil-core dependency to fe-foundation pom.xml.
Tests remain in fe-core as they depend on GsonUtils for serialization tests.
