+++ fe/fe-core/src/test/java/org/apache/doris/foundation/util/ConcurrentLong2ObjectHashMapTest.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.doris.common;
+package org.apache.doris.foundation.util;
 
 import org.apache.doris.persist.gson.GsonUtils;
